requirements.txt
requests>=2.31.0
scikit-learn>=1.3.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
